# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "York academy",
        "description": "York academy",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/"
        },
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/venue/{id}": {
            "get": {
                "tags": [
                    "Venues"
                ],
                "summary": "Get a venue by ID",
                "description": "Retrieve details of a specific venue by its ID",
                "operationId": "getVenueById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the venue",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venue details retrieved successfully"
                    },
                    "404": {
                        "description": "Venue not found"
                    }
                }
            }
        }
    },
...
}